@@ -87,16 +88,16 @@
     "wait-on": "^7.0.1"
   },
   "dependencies": {
-    "@angular/animations": "17.3.2",
-    "@angular/common": "17.3.2",
-    "@angular/compiler": "17.3.2",
-    "@angular/core": "17.3.2",
-    "@angular/forms": "17.3.2",
-    "@angular/platform-browser": "17.3.2",
-    "@angular/platform-browser-dynamic": "17.3.2",
-    "@angular/pwa": "17.3.2",
-    "@angular/router": "17.3.2",
-    "@angular/service-worker": "17.3.2",
+    "@angular/animations": "18.0.7",
+    "@angular/common": "18.0.7",
+    "@angular/compiler": "18.0.7",
+    "@angular/core": "18.0.7",
+    "@angular/forms": "18.0.7",
+    "@angular/platform-browser": "18.0.7",
+    "@angular/platform-browser-dynamic": "18.0.7",
+    "@angular/pwa": "18.0.7",
+    "@angular/router": "18.0.7",
+    "@angular/service-worker": "18.0.7",
     "@apollo/client": "^3.8.10",
     "@automapper/classes": "^8.8.1",
     "@automapper/core": "^8.8.1",
