@@ -1,5 +1,5 @@
-import { TraceClassDecorator } from '@sentry/angular-ivy';
+import { TraceClass } from '@sentry/angular';
 
 export function TraceComponent(): ClassDecorator {
-	return TraceClassDecorator();
+	return TraceClass();
 }
